requirements.txt
# Azure MCP Server Dependencies
azure-monitor-query==1.2.0
azure-search-documents==11.4.0
pydantic>=2.0
semantic-kernel>=1.0
azure-identity
azure-ai-projects
azure-identity==1.14.0
aiohttp
httpx
fastapi
uvicorn